+++ MiniIndex/webpack.config.js
@@ -1,0 +1,32 @@
+const glob = require('glob');
+const MiniCssExtractPlugin = require("mini-css-extract-plugin");
+
+module.exports = {
+    entry: {
+        site: './wwwroot/js/site.js',
+        bootstrap_js: './wwwroot/js/bootstrapjs.js',
+        validation: './wwwroot/js/validation.js',
+        tags: './wwwroot/js/tags-selectized.js',
+        stars: './wwwroot/js/stars.js'
+    },
+    output: {
+        filename: '[name].entry.js',
+        path: __dirname + '/wwwroot/dist'
+    },
+    devtool: 'source-map',
+    mode: 'development',
+    module: {
+        rules: [
+            { test: /\.css$/, use: [{ loader: MiniCssExtractPlugin.loader }, "css-loader"] },
+            { test: /\.eot(\?v=\d+\.\d+\.\d+)?$/, loader: "file-loader" },
+            { test: /\.(woff|woff2)$/, loader: "url-loader?prefix=font/&limit=5000" },
+            { test: /\.ttf(\?v=\d+\.\d+\.\d+)?$/, loader: "url-loader?limit=10000&mimetype=application/octet-stream" },
+            { test: /\.svg(\?v=\d+\.\d+\.\d+)?$/, loader: "url-loader?limit=10000&mimetype=image/svg+xml" }
+        ]
+    },
+    plugins: [
+        new MiniCssExtractPlugin({
+            filename: "[name].css"
+        })
+    ]
+};
