+++ MiniIndex/wwwroot/js/bootstrapjs.js
@@ -1,0 +1,1 @@
+﻿import 'bootstrap'
