+++ MiniIndex/wwwroot/js/validation.js
@@ -1,0 +1,3 @@
+﻿import 'jquery'
+import 'jquery-validation'
+import 'jquery-validation-unobtrusive'
